refactor(cli): organize server commands
