+++ app/src/cli/commands/server/handler.rs
@@ -1,6 +1,6 @@
 use crate::cli::output;
+use super::{ServerCommand, ServerLaunchArgs};
 use crate::{
-  cli::args::{ServerCommand, ServerLaunchArgs},
   config::{ServerConfig, ServerOverrides, resolve_data_dir, sqlite_url},
   constants::config::DATABASE_FILENAME,
   daemon::ManagedDaemonState,
