+++ app/src/cli/commands/server/mod.rs
@@ -1,3 +1,6 @@
+mod args;
 mod handler;
 
+pub use args::{ServerCommand, ServerLaunchArgs, ServerStartArgs};
+pub(crate) use args::HELP;
 pub(super) use handler::execute;
